update manage account
